Commit 2: Electric Boogaloo
Created an IPython Notebook to keep track of progress and data
exploration. Optimized initial filtering of LCA data (it’s not actually
H1B applications - important distinction). Combined “Clean_Up.py” and
“Initial_Filter.py” into one file and removed code and comments related
to exploring the data. All data exploration will be done in the
notebooks. The .py files will are there to quickly generate the end
product of each step (in this case, a filtered and prepped cvs file).